# Improve performance populate_for_version