--- a/changelogs/unreleased/improve-performance-populate-for-version.yml
+++ b/changelogs/unreleased/improve-performance-populate-for-version.yml
@@ -1,0 +1,4 @@
+---
+description: "Improve the performance of the `populate_for_version` query."
+change-type: patch
+destination-branches: [master, iso8]